Refactor extract utility functions
